requirements.txt
python-telegram-bot==22.3
openai>=1.40.0
httpx>=0.27.0
python-dotenv>=1.0.1
md2pdf